Began artwork on vehicles, particle optimizations ...
- Vehicle artwork has started and sprites are in the game
- Vehicle objects added for each vehicle, going off a parent obj_vehicleP for code
- There are now two different particle objects: static and dynamic:
Static: Static particle objects are objects who have a single emitter for an entire particle system. They only need to be created at the start of the room and the game will handle the rest It is moved around manually and set to burst to use it. This is much more efficient than dynamic particle systems.
Dynamic: dynamic particles are particles that have an emitter for each instance of themselves. They are used when special events are used for it that cannot be done just with a single particle emitter. This is less efficient than static particle systems.
- blood spurts and zombie gore are static particle objects, more of those particles can be handled with a lot less of a performace hit.
- the weapons.json file is now an included file within Gamemaker, that way we don't have to create it manually within the local appdata folder every time the game is launched.